Implement starttls extension
Enable using TLS on peer connections via a starttls extension message.
This is a first step towards implementing the proposed persistent
credits enhancement described here:

https://github.com/ssiloti/bep-persistent-credit/blob/master/persistent-credit.rst

TLS is used as part of the proposal to provide secure authentication of
peer identity.

The implementation uses very different approaches for the read and write
paths within the tr_peerIo object. Reads are implemented by using a
custom BIO to feed the TLS object with data from the underlying TCP/UTP
connection. The decrypted data is then read into the peerIo's inbuf.
This has the downside of requiring separate implementations for TCP and
UTP. The other option of using a custom BIO to feed data from the inbuf
to the TLS object then reading directly into the buffers provided by the
peerIO's clients was discarded because the peerIo interface implicitly
assumes a one-to-one relationship between the size of the inbuf and the
number of bytes available for reading. Because TLS incurs overhead for
record headers and padding this relationship would no longer hold.

Writes are handled by using a custom BIO to write the encrypted TLS
stream into the peerIo's outbuf. A second buffer is required to store
unencrypted data during a renegotiation. The option of using a BIO to
write directly to the underlying TCP/UTP connection was discarded
because the UTP API requires that the peerIo know upfront how much data
needs to be written which it cannot know without first processing the
output data through the TLS object. Implementing the write path in this
manner would require the use of an extra buffer, and incur an extra
copy, at all times instead of only during renegotiation.